refactor(rules): Base setNextTime call only on noWait parameter
This commit refactors the `executeAction` function to make the call to `setNextTime` conditional only on the `noWait` parameter.

Previously, the call was also conditional on the service name not being "slack". This restriction has been removed to make the behavior more consistent and predictable. Now, `setNextTime` will be called for all services whenever `noWait` is false, ensuring that the next run time is always updated unless explicitly disabled.